% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conwy-data.R
\docType{data}
\name{ConwyLU}
\alias{ConwyLU}
\title{Current land use map (heavily aggregated)}
\description{
An object of class \href{https://cran.r-project.org/package=raster}{\code{RasterLayer}}, sourced from \emph{extdata/ConwyLU.tif}. 
A simplified version of the current land use map from the Conwy catchment (Wales, UK). 
It includes three classes: arable (raster value 3), forest (2), other (1).
}

