% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conwy-data.R
\docType{data}
\name{ConwySlope}
\alias{ConwySlope}
\title{Slope raster for Conwy catchment.}
\description{
An object of class RasterLayer, sourced from \emph{extdata/ConwySlope.tif}. A raster of slope derived from a digital elevation model at 50 meters resolution, units are degrees.
}

