% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conwy-data.R
\docType{data}
\name{ConwyStatus}
\alias{ConwyStatus}
\title{Ownership/legal status raster for Conwy catchment (dummy data).}
\description{
An object of class RasterLayer, sourced from \emph{extdata/ConwyStatus.tif}. The land ownership type (dummy data), divided into three possible classes: public (raster value 4), private (3), protected (1).
}

