% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conwy-data.R
\docType{data}
\name{LUclasses}
\alias{LUclasses}
\title{Lookup list, linking input spatial data and the Bayesian network}
\description{
A list with the classification of input spatial data (its corresponding states and values). 
The list is formatted accordingly to bnspatial functions requirement and as returned by 
functions \code{\link{importClasses}} and \code{\link{setClasses}}.
}

